requirements.txt
 pandas>=2.0.0
python-dotenv>=1.0.0
google-genai>=1.47.0
fastapi>=0.128.0
uvicorn>=0.39.0
python-multipart>=0.0.20
pytest>=8.0.0
